@antonfirsov 2023-08-25
Yeah, according to the article we should divide by `255` to get the intensity bin index instead of defining 255x more buckets.

```C#
int currentIntensity = (int)MathF.Round((sourceBlue + sourceGreen + sourceRed) / 3F * this.levels / 255f);
```

But with this tests are failing and the output looks like a simple blur. Unfortunately I cannot go deeper understanding right now.